requirements.txt
streamlit>=1.28.0
streamlit-plotly-events>=0.0.5
pandas>=2.0.0
numpy>=1.24.0
duckdb>=0.9.0
plotly>=5.18.0
rapidfuzz>=3.5.0
scikit-learn>=1.3.0
umap-learn>=0.5.5
hdbscan>=0.8.33
openai>=1.12.0
python-dotenv>=1.0.0
pyarrow>=14.0.0
pytest>=7.0.0
requests>=2.28.0
boto3>=1.28.0
soccerdata>=1.8.0
lxml>=4.9.0
